Refactor: Migrate API definition to workbench-lite schema
