+++ src/index.ts
@@ -3,12 +3,6 @@ export function helloWorld(name?: string): string {
   return `Hello, ${who}!`;
 }
 
-export { Api } from "./api/api.js";
-export * from "./api/registry.js";
-export * from "./api/snapshot.js";
-export * from "./api/api-host.js";
-export * from "./api/doc-renderer.js";
-
 export * from "./agnet.js";
 export * from "./providers.js";
 export * from "./agent-mdx.js";
@@ -22,4 +16,6 @@ export * from "./workbench-lite-sse.js";
 export * as workbenchLight from "./workbench-light.js";
 
 export * from "./app.js";
+export * from "./api-snapshot.js";
+export * from "./api-doc-renderer.js";
